requirements.txt
Django==5.0.6
django_filter==2.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
phonenumbers==8.13.40
pyotp==2.9.0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.7
Requests==2.32.3
django-redis==5.4.0
Pillow==10.1.0
gunicorn==21.2.0
django-cors-headers==3.13.0
drf-yasg==1.21.7